requirements.txt
annotated-doc==0.0.4
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
idna==3.11
markdown-it-py==4.0.0
mdurl==0.1.2
Pygments==2.19.2
python-dotenv==1.2.2
requests==2.33.0
rich==14.3.3
shellingham==1.5.4
typer==0.24.1
urllib3==2.6.3
